add socket event in myauction page
